backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
groq==0.4.1
asyncpg==0.29.0
openai==1.3.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
psycopg2-binary==2.9.7
